CRUD do Projeto e Equipe
Vou fazer o crud do canvas agr. Por favor n mexer

+++ StartupManager/ModelUsuario.cs
@@ -48,7 +48,7 @@ namespace StartupManager
             return dt;
         }
 
-        public Usuario BuscaId(int id)
+        public Usuario BuscaId(Int64 id)
         {
             Usuario u = null;
             try
